requirements.txt
asgiref==3.2.3
astroid==2.0.4
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
future==0.18.2
gunicorn==20.0.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
liffpy==0.0.1a3
line-bot-sdk==1.15.0
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.1.1
pytz==2019.3
requests==2.22.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
wrapt==1.10.11
